+++ test/test_helper.exs
@@ -1,3 +1,1 @@
 ExUnit.start()
-
-Code.require_file("support/rustq_package_case.ex", __DIR__)